refactors widgets into their own files, starts refactoring database connection
